MNT: separate fins classes
