+++ rocketpy/rocket/aero_surface/fins/__init__.py
@@ -1,0 +1,3 @@
+from rocketpy.rocket.aero_surface.fins.fins import Fins
+from rocketpy.rocket.aero_surface.fins.trapezoidal_fins import TrapezoidalFins
+from rocketpy.rocket.aero_surface.fins.elliptical_fins import EllipticalFins
